requirements.txt
python-telegram-bot
requests
pyyaml
